Minor Key/Value pairs json refactor (#384)
* Just some refactoring to move things around, functionally equivalent.

* silliness
